requirement.txt
aiohttp==3.8.4
aiosignal==1.3.1
ale-py==0.8.1
altair==4.1.0
ansible-base==2.10.8
apache-libcloud==3.2.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
async-timeout==4.0.2
attrs==23.1.0
autonomous-learning-library==0.6.2
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
box2d-py==2.3.5
Brlapi==0.8.3
cachetools==4.2.4
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
cloudpickle==3.0.0
cmake==3.28.1
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
datasets==2.16.0
dateparser==1.2.0
dbus-python==1.2.18
defer==1.0.6
dill==0.3.7
distro==1.7.0
distro-info===1.1build1
dnspython==2.1.0
docker==6.0.1
docker-pycreds==0.4.0
duplicity==0.8.21
entrypoints==0.4
fasteners==0.14.1
filelock==3.13.1
fonttools==4.47.0
frozenlist==1.3.3
fsspec==2023.12.2
future==0.18.2
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
google-api-core==1.31.5
google-api-python-client==2.7.0
google-auth==1.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
gym==0.26.0
gym-notices==0.0.8
httmock==1.4.0
httplib2==0.20.2
huggingface-hub==0.20.1
idna==3.3
importlib-metadata==4.6.4
importlib-resources==6.1.1
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lit==17.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.8.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mysql-connector-python==8.1.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.2.1
nltk==3.8.1
ntlm-auth==1.4.0
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
openai==0.27.8
opencv-python==3.4.18.65
packaging==21.3
pandas==2.2.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
prometheus-api-client==0.5.5
protobuf==4.21.12
psutil==5.9.7
psycopg2==2.9.7
ptyprocess==0.7.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybullet==3.2.6
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
pydeck==0.8.1b0
pygame==2.1.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pykerberos==1.1.14
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.5.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.9.0.post0
python-debian===0.1.43ubuntu1
pytz==2022.1
pywinrm==0.3.0
pyxdg==0.27
PyYAML==5.4.1
referencing==0.32.0
regex==2023.12.25
reportlab==3.6.8
requests==2.29.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
rpds-py==0.15.2
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
SecretStorage==3.3.1
selinux==3.3
semantic-version==2.10.0
semver==3.0.2
sentence-transformers==1.1.1
sentencepiece==0.1.99
sentry-sdk==1.39.1
seqeval==1.2.2
setproctitle==1.3.3
setuptools-rust==1.8.1
simplejson==3.17.6
simpletransformers==0.61.4
six==1.16.0
smmap==5.0.1
soupsieve==2.5
ssh-import-id==5.11
stable-baselines3==1.0
streamlit==1.9.0
swig==4.1.1.post1
sympy==1.12
systemd-python==234
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.4
tqdm==4.65.0
transformers==4.30.0
triton==2.0.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.18
usb-creator==0.3.7
validators==0.22.0
wadllib==1.3.6
wandb==0.16.1
warc==0.2.1
watchdog==3.0.0
websocket-client==1.5.1
xdg==5
xkit==0.0.0
xmltodict==0.12.0
xxhash==3.4.1
yarl==1.9.2
zipp==1.0.0
